refactor(operator_commands_dashboard): extract goals API to dedicated module
Move goals/seed_goals/add_goal/remove_goal/update_goal_status/
promote_backlog_item/demote_goal HTTP handlers from routes.rs into a new
sibling goals.rs module. Promotes resolve_state_root to pub(crate) so
goals can share it.

- routes.rs: 5488 -> 5123
- goals.rs (NEW, 377)

Refs #1266
